chore(utils): Remove excess simple type casts.

+++ zellij-utils/src/pane_size.rs
@@ -205,7 +205,7 @@ impl Eq for Constraint {}
 
 impl PaneGeom {
     pub fn contains(&self, point: &Position) -> bool {
-        let col = point.column.0 as usize;
+        let col = point.column.0;
         let row = point.line.0 as usize;
         self.x <= col
             && col < self.x + self.cols.as_usize()
